+++ src/main/java/org/embulk/util/guess/CharsetRecog_2022.java
@@ -1,10 +1,17 @@
 /*
+ * This file is based on a copy from ICU4J 57.2 with modification on :
+ * - moving its Java package to org.embulk.util.guess.
+ *
+ * It is licensed under the ICU License, which is bundled as LICENSE.icu4j.
+ */
+
+/*
 *******************************************************************************
 * Copyright (C) 2005 - 2012, International Business Machines Corporation and  *
 * others. All Rights Reserved.                                                *
 *******************************************************************************
 */
-package com.ibm.icu.text;
+package org.embulk.util.guess;
 
 /**
  *  class CharsetRecog_2022  part of the ICU charset detection imlementation.
