+++ src/main/java/org/embulk/util/guess/CharsetRecog_mbcs.java
@@ -1,11 +1,18 @@
 /*
+ * This file is based on a copy from ICU4J 57.2 with modification on :
+ * - moving its Java package to org.embulk.util.guess.
+ *
+ * It is licensed under the ICU License, which is bundled as LICENSE.icu4j.
+ */
+
+/*
  ****************************************************************************
  * Copyright (C) 2005-2012, International Business Machines Corporation and *
  * others. All Rights Reserved.                                             *
  ****************************************************************************
  *
  */
-package com.ibm.icu.text;
+package org.embulk.util.guess;
 
 import java.util.Arrays;
