+++ src/main/java/org/embulk/util/guess/CharsetRecog_sbcs.java
@@ -1,4 +1,11 @@
 /*
+ * This file is based on a copy from ICU4J 57.2 with modification on :
+ * - moving its Java package to org.embulk.util.guess.
+ *
+ * It is licensed under the ICU License, which is bundled as LICENSE.icu4j.
+ */
+
+/*
  ****************************************************************************
  * Copyright (C) 2005-2013, International Business Machines Corporation and *
  * others. All Rights Reserved.                                             *
@@ -6,7 +13,7 @@
  *
  */
 
-package com.ibm.icu.text;
+package org.embulk.util.guess;
 
 /**
  * This class recognizes single-byte encodings. Because the encoding scheme is so
